Refactor: id 이름 수정 및 파일 이름 변경
