+++ lib/custom_map/components/marker/google_user_marker.dart
@@ -1,4 +1,4 @@
-import 'package:frontend/custom_map/model/type.dart';
+import 'package:frontend/custom_map/model/marker_static_info_model.dart';
 import 'package:frontend/custom_map/components/marker/user_marker_icon.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
