+++ lib/custom_map/components/test_button/create_init_marker_button.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:frontend/common/consts/data.dart';
-import 'package:frontend/custom_map/model/type.dart';
+import 'package:frontend/custom_map/model/marker_static_info_model.dart';
 import 'package:frontend/custom_map/components/marker/google_user_marker.dart';
 import 'package:frontend/custom_map/components/marker/user_marker_icon.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
