+++ lib/custom_map/model/friend_info_model.dart
@@ -27,16 +27,16 @@ class FriendLocationAndBattery {
 }
 
 class FriendInfoModel {
-  late String userNameTag;
+  late String userId;
   late LiveInfoModel liveInfo;
   FriendInfoModel({
-    required this.userNameTag,
+    required this.userId,
     required this.liveInfo,
   });
 
   factory FriendInfoModel.fromJson(Map<String, dynamic> json) {
     return FriendInfoModel(
-      userNameTag: json.keys.first,
+      userId: json.keys.first,
       liveInfo: LiveInfoModel.fromJson(json.values.first),
     );
   }
