+++ lib/custom_map/model/marker_static_info_model.dart
